requirements.txt
Django==5.0
Pillow==10.1.0
django-crispy-forms==2.1
python-dotenv==1.0.0
stripe==7.10.0
